src/backend/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
requests==2.31.0
opentelemetry-api==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-exporter-otlp==1.18.0
opentelemetry-instrumentation-flask==0.39b0
opentelemetry-instrumentation-requests==0.39b0
opentelemetry-instrumentation-psycopg2==0.39b0
python-dotenv==1.0.0
pytest==7.4.0
pytest-flask==1.2.0